same change in java code
