+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/extension/model/codemodel/Operation.java
@@ -34,7 +34,6 @@ public class Operation extends Metadata {
     private ConvenienceApi convenienceApi;
     private Boolean generateProtocolApi;
     private Boolean internalApi;
-    private String crossLanguageDefinitionId;
     // internal
     private OperationGroup operationGroup;
 
@@ -404,24 +403,6 @@ public class Operation extends Metadata {
         this.internalApi = internalApi;
     }
 
-    /**
-     * Gets the cross-language definition ID.
-     *
-     * @return The cross-language definition ID.
-     */
-    public String getCrossLanguageDefinitionId() {
-        return crossLanguageDefinitionId;
-    }
-
-    /**
-     * Sets the cross-language definition ID.
-     *
-     * @param crossLanguageDefinitionId The cross-language definition ID.
-     */
-    public void setCrossLanguageDefinitionId(String crossLanguageDefinitionId) {
-        this.crossLanguageDefinitionId = crossLanguageDefinitionId;
-    }
-
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         return super.writeParentProperties(jsonWriter.writeStartObject()).writeStringField("operationId", operationId)
@@ -443,7 +424,6 @@ public class Operation extends Metadata {
             .writeJsonField("convenienceApi", convenienceApi)
             .writeBooleanField("generateProtocolApi", generateProtocolApi)
             .writeBooleanField("internalApi", internalApi)
-            .writeStringField("crossLanguageDefinitionId", crossLanguageDefinitionId)
             .writeEndObject();
     }
 
@@ -498,8 +478,6 @@ public class Operation extends Metadata {
                 operation.generateProtocolApi = reader.getBoolean();
             } else if ("internalApi".equals(fieldName)) {
                 operation.internalApi = reader.getBoolean();
-            } else if ("crossLanguageDefinitionId".equals(fieldName)) {
-                operation.crossLanguageDefinitionId = reader.getString();
             } else {
                 reader.skipChildren();
             }
